Move create_vellum_client to vellum.utils and reexport from vellum.workflows.vellum_client
- Moved create_vellum_client, create_vellum_environment, and _resolve_env helper functions to src/vellum/utils/vellum_client.py
- Updated src/vellum/workflows/vellum_client.py to reexport these functions for backward compatibility
- Also reexport Vellum class to maintain existing imports
- Updated src/vellum/utils/__init__.py to export create_vellum_client
