+++ src/vellum/utils/__init__.py
@@ -1,0 +1,3 @@
+from vellum.utils.vellum_client import create_vellum_client
+
+__all__ = ["create_vellum_client"]
